+++ relay-conventions/src/lib.rs
@@ -235,7 +235,7 @@ mod tests {
         insta::assert_debug_snapshot!(info, @r###"
         AttributeInfo {
             write_behavior: CurrentName,
-            pii: Maybe,
+            pii: True,
             aliases: [
                 "params.<key>",
             ],