+++ tests/integration/test_spansv2.py
@@ -65,6 +65,9 @@ def test_spansv2_basic(
     project_config["config"].update(
         {"retentions": {"span": {"standard": 42, "downsampled": 1337}}}
     )
+    project_config["config"].setdefault("features", []).extend(
+        ["organizations:relay-generate-billing-outcome"]
+    )
 
     relay = relay(relay_with_processing(options=TEST_CONFIG), options=TEST_CONFIG)
 
@@ -200,6 +203,7 @@ def test_spansv2_basic(
             "tags": {
                 "was_transaction": "false",
                 "is_segment": "true",
+                "billing_outcome_emitted": "true",
             },
             "timestamp": time_within_delta(),
             "type": "c",
@@ -207,8 +211,30 @@ def test_spansv2_basic(
         },
     ]
 
+    num_messages = 2
     if relay_emits_accepted_outcome:
-        assert outcomes_consumer.get_aggregated_outcomes() == [
+        num_messages = 3
+
+    outcomes = outcomes_consumer.get_aggregated_outcomes(n=num_messages)
+
+    if relay_emits_accepted_outcome:
+        assert outcomes == [
+            {
+                "category": DataCategory.TRANSACTION.value,
+                "key_id": 123,
+                "org_id": 1,
+                "outcome": 0,
+                "project_id": 42,
+                "quantity": 1,
+            },
+            {
+                "category": DataCategory.SPAN.value,
+                "key_id": 123,
+                "org_id": 1,
+                "outcome": 0,
+                "project_id": 42,
+                "quantity": 1,
+            },
             {
                 "category": DataCategory.SPAN_INDEXED.value,
                 "key_id": 123,
@@ -216,7 +242,27 @@ def test_spansv2_basic(
                 "outcome": 0,
                 "project_id": 42,
                 "quantity": 1,
-            }
+            },
+        ]
+
+    else:
+        assert outcomes == [
+            {
+                "category": DataCategory.TRANSACTION.value,
+                "key_id": 123,
+                "org_id": 1,
+                "outcome": 0,
+                "project_id": 42,
+                "quantity": 1,
+            },
+            {
+                "category": DataCategory.SPAN.value,
+                "key_id": 123,
+                "org_id": 1,
+                "outcome": 0,
+                "project_id": 42,
+                "quantity": 1,
+            },
         ]
 
 
@@ -236,6 +282,9 @@ def test_spansv2_trimming_basic(
 
     project_id = 42
     project_config = mini_sentry.add_full_project_config(project_id)
+    project_config["config"].setdefault("features", []).append(
+        "organizations:relay-generate-billing-outcome"
+    )
     project_config["config"].update(
         {
             "retentions": {"span": {"standard": 42, "downsampled": 1337}},
@@ -409,6 +458,7 @@ def test_spansv2_trimming_basic(
             "tags": {
                 "was_transaction": "false",
                 "is_segment": "true",
+                "billing_outcome_emitted": "true",
             },
             "timestamp": time_within_delta(),
             "type": "c",
@@ -635,10 +685,16 @@ def test_spansv2_ds_sampled(
 
     project_id = 42
     project_config = mini_sentry.add_full_project_config(project_id)
+    project_config["config"].setdefault("features", []).extend(
+        ["organizations:relay-generate-billing-outcome"]
+    )
     add_sampling_config(project_config, sample_rate=0.0, rule_type="trace")
 
     sampling_project_id = 43
     sampling_config = mini_sentry.add_basic_project_config(sampling_project_id)
+    sampling_config["config"].setdefault("features", []).extend(
+        ["organizations:relay-generate-billing-outcome"]
+    )
     sampling_config["organizationId"] = project_config["organizationId"]
     add_sampling_config(sampling_config, sample_rate=0.9, rule_type="trace")
 
@@ -724,6 +780,7 @@ def test_spansv2_ds_sampled(
             "retention_days": 90,
             "tags": {
                 "is_segment": "false",
+                "billing_outcome_emitted": "true",
             },
             "timestamp": time_within_delta(),
             "type": "c",
@@ -738,6 +795,7 @@ def test_spansv2_ds_sampled(
             "tags": {
                 "was_transaction": "false",
                 "is_segment": "true",
+                "billing_outcome_emitted": "true",
             },
             "timestamp": time_within_delta(),
             "type": "c",
@@ -745,7 +803,25 @@ def test_spansv2_ds_sampled(
         },
     ]
 
-    assert outcomes_consumer.get_aggregated_outcomes(n=1) == [
+    outcomes = outcomes_consumer.get_aggregated_outcomes(n=3)
+
+    assert outcomes == [
+        {
+            "category": DataCategory.TRANSACTION.value,
+            "key_id": 123,
+            "org_id": 1,
+            "outcome": 0,
+            "project_id": 42,
+            "quantity": 1,
+        },
+        {
+            "category": DataCategory.SPAN.value,
+            "key_id": 123,
+            "org_id": 1,
+            "outcome": 0,
+            "project_id": 42,
+            "quantity": 2,
+        },
         {
             "category": DataCategory.SPAN_INDEXED.value,
             "key_id": 123,
@@ -753,7 +829,7 @@ def test_spansv2_ds_sampled(
             "outcome": 0,
             "project_id": 42,
             "quantity": 2,
-        }
+        },
     ]
 
 
@@ -776,10 +852,18 @@ def test_spansv2_ds_root_in_different_org(
 
     project_id = 42
     project_config = mini_sentry.add_full_project_config(project_id)
+    project_config["config"].setdefault("features", []).extend(
+        ["organizations:relay-generate-billing-outcome"]
+    )
+
     add_sampling_config(project_config, sample_rate=0.0, rule_type="trace")
 
     sampling_project_id = 43
     sampling_config = mini_sentry.add_basic_project_config(sampling_project_id)
+    sampling_config["config"].setdefault("features", []).extend(
+        ["organizations:relay-generate-billing-outcome"]
+    )
+
     sampling_config["organizationId"] = 99
     add_sampling_config(sampling_config, sample_rate=1.0, rule_type="trace")
 
@@ -829,6 +913,7 @@ def test_spansv2_ds_root_in_different_org(
             "retention_days": 90,
             "tags": {
                 "is_segment": "false",
+                "billing_outcome_emitted": "true",
             },
             "timestamp": time_within_delta(),
             "type": "c",
@@ -836,16 +921,27 @@ def test_spansv2_ds_root_in_different_org(
         },
     ]
 
-    assert outcomes_consumer.get_outcome() == {
-        "category": DataCategory.SPAN_INDEXED.value,
-        "key_id": 123,
-        "org_id": 1,
-        "outcome": 1,
-        "project_id": 42,
-        "quantity": 1,
-        "reason": "Sampled:0",
-        "timestamp": time_within_delta(),
-    }
+    assert outcomes_consumer.get_outcomes(n=2) == [
+        {
+            "category": DataCategory.SPAN_INDEXED.value,
+            "key_id": 123,
+            "org_id": 1,
+            "outcome": 1,
+            "project_id": 42,
+            "quantity": 1,
+            "reason": "Sampled:0",
+            "timestamp": time_within_delta(),
+        },
+        {
+            "category": DataCategory.SPAN.value,
+            "key_id": 123,
+            "org_id": 1,
+            "outcome": 0,
+            "project_id": 42,
+            "quantity": 1,
+            "timestamp": time_within_delta(),
+        },
+    ]
 
     spans_consumer.assert_empty()