+++ CHANGELOG.md
@@ -6,6 +6,11 @@
 
 - Don't reject attributes that don't have values, but do have metadata. ([#6098](https://github.com/getsentry/relay/pull/6098))
 
+**Internal**:
+
+- Require timestamps and verification in auth signatures. ([#6069](https://github.com/getsentry/relay/pull/6069))
+- Have relay generate metric billing outcomes. ([#6066](https://github.com/getsentry/relay/pull/6066))
+
 ## 26.6.0
 
 **Features**: